requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
streamlit>=1.28.0
qrcode>=7.4.2
cryptography>=41.0.0
Pillow>=10.0.0
openpyxl>=3.1.0
bcrypt>=4.0.0
PyJWT>=2.8.0
